+++ common/src/main/java/com/infernalstudios/infernalexp/compat/TerraBlenderCompat.java
@@ -2,13 +2,10 @@ package com.infernalstudios.infernalexp.compat;
 
 import com.infernalstudios.infernalexp.IEConstants;
 import com.infernalstudios.infernalexp.world.surface.ModSurfaceRules;
-import net.minecraft.resources.ResourceLocation;
-import terrablender.api.Regions;
 import terrablender.api.SurfaceRuleManager;
 
 public class TerraBlenderCompat {
     public static void register() {
-        Regions.register(new TerraBlenderRegion(ResourceLocation.fromNamespaceAndPath(IEConstants.MOD_ID, "nether"), 4));
         SurfaceRuleManager.addSurfaceRules(SurfaceRuleManager.RuleCategory.NETHER, IEConstants.MOD_ID, ModSurfaceRules.addNetherSurfaceRulesWithBedrock());
     }
 }