+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.13] - 2026-08-18
+
+### Changed
+
+- Reworked biome generation.
+- Lithostitched is now a required dependency.
+
+### Fixed
+
+- Fixed Amplified Nether incompatibility.
+
 ## [0.3.12] - 2026-07-23
 
 ### Added